* device/include/mcs51/reg51.h: fall back to 8052.h * doc/sdccman.lyx: pragma exclude bits * src/mcs51/gen.c (genFunction, genEndFunction): applied patch 193 for   pragma exclude bits * src/port.h: maxextalign warning fixed * src/SDCCglobl.h: define true and false for MSVC * src/stm8/gen.c (genCmp): fixed warning * src/stm8/ralloc2.cc (Ainst_ok): fixed warning
git-svn-id: http://svn.code.sf.net/p/sdcc/code/trunk@8878 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ sdcc/src/stm8/ralloc2.cc
@@ -135,7 +135,7 @@ static bool Ainst_ok(const assignment &a, unsigned short int i, const G_t &G, co
         return(false);
 
       // Register pushes do not disturb a.
-      for (int i = 0; i < getSize(operandType(IC_LEFT(ic)));)
+      for (unsigned short i = 0; i < getSize(operandType(IC_LEFT(ic)));)
         {
           if(operand_in_reg(left, REG_A, ia, i, G))
             i++;
